--- a/model/.gitignore
+++ b/model/.gitignore
@@ -1,0 +1,2 @@
+*.tflite
+text2vec-base-chinese/*