--- a/model/clone_model_repository.bat
+++ b/model/clone_model_repository.bat
@@ -1,0 +1,2 @@
+git lfs install
+git clone https://huggingface.co/shibing624/text2vec-base-chinese