--- a/model/clone_model_repository.bat
+++ b/model/clone_model_repository.bat
@@ -1,2 +1,0 @@
-git lfs install
-git clone https://huggingface.co/shibing624/text2vec-base-chinese