--- a/model/download_bert_embedder.bat
+++ b/model/download_bert_embedder.bat
@@ -1,0 +1,1 @@
+curl -o embedder.tflite https://storage.googleapis.com/mediapipe-models/text_embedder/bert_embedder/float32/1/bert_embedder.tflite