… docker-compose.yaml
… flask4modelcache_demo.py
… model/text2vec-base-chinese/logs.txt
… modelcache/__init__.py
… modelcache/adapter/adapter.py
… modelcache/adapter/adapter_insert.py
… modelcache/adapter/adapter_query.py
… modelcache/adapter/adapter_register.py
… modelcache/adapter/adapter_remove.py
… modelcache/cache.py
… modelcache/config.py
… modelcache/config/milvus_config.ini
… modelcache/config/mysql_config.ini
… modelcache/core.py
… modelcache/embedding/__init__.py
… modelcache/embedding/base.py
… modelcache/embedding/embedding_dispatcher.py
… modelcache/embedding/huggingface.py
… modelcache/embedding/mpnet_base.py
… modelcache/embedding/string_text.py
… modelcache/manager/__init__.py
… modelcache/manager/data_manager.py
… modelcache/manager/eviction/__init__.py
… modelcache/manager/eviction/arc_cache.py
… modelcache/manager/eviction/base.py
… modelcache/manager/eviction/manager.py
… modelcache/manager/eviction/memory_cache.py
… modelcache/manager/eviction/wtinylfu_cache.py
… modelcache/manager/factory.py
… modelcache/manager/object_data/__init__.py