Hello.

We're 4 university students who have been working on improving this project for an entire semester
as part of a course called 'Caching in LLMs'. The improvements we've added target nearly every part of the system.

We've added significant performance improvements, long-overdue refactoring of the code and new features.

### The problem domain:

The current ModelCache project is missing several critical features and is not optimized to yield the best performance possible.

1) There is no in-memory cache for saved queries (other then mysql's cache)

2) The model used for embedding is not optimal for the task of semantic search

3) The system does not support bulk operations

4) The system is based solely on a RESTful API, which has the overhead of the handshake and connection establishment for each and every query

5) Database access logic is inefficient and not scalable

6) The embedding process is completely single-threaded due to python's GIL and is incapable of serving multiple clients, while not using the CPU to its fullest potential.

7) The project is currently running on the old Python 10 version

Our changes aim to address all these problems.


### How the improvements have been tested:

We created a kotlin-based test suite project to benchmark and test the ModelCache system.
The testing project can be found [in this repository](https://github.com/Yuval-Roth/modelcache_testing)
The dataset we used to test the system [can be found here](https://qa.mpi-inf.mpg.de/comqa/).

From the dataset, we picked entries with at least 2 questions (out of which we took 2 at random) and 1 answer.
This resulted in the [following file that contains the final test data with 1666 entries](https://github.com/Yuval-Roth/ModelCache_Testing/blob/main/src/main/resources/comqa_filtered.json)
We then tested the hit ratio of the cache, its query throughput and its latency.

#### The test:
The test revolves around picking one question out of each pair and inserting them all to the cache
and then querying the other questions that weren't inserted.

#### The noteworthy metrics we tested are:
* Insertion time
* Querying time
* Hit ratio
*  Expected query hit ratio (the ratio of queries that hit another question with the same answer)
* Query throughput
* Mean query latency
* p95 query latency
* p99 percentile query latency

### Performance benchmarks

#### System specifications:
* CPU: Intel Core i7-14700KF
* Ram: 32GB, DDR5, 6400MHz, CL32
* Storage: NVMe SSD
* Operating system: Windows 11


The system before our changes is limited to 1 client and 1 worker.
We can see massive performance gains on every metric, when comparing the same setup of 1 client and 1 worker

#### before any changes were made - 1 client, 1 worker, flask4modelcache
```
| Metric                        | Value                      |
|-------------------------------|----------------------------|
| Insertion time                | 80529 ms                   |
| Querying time                 | 79763 ms                   |
| Hit ratio                     | 1036 / 1668 (62.11031%)    |
| Expected query hit ratio      | 911 / 1036 (87.934364%)    |
| Query throughput              | 20.911951 queries/second   |
| Mean query latency            | 47.793766 ms               |
| p95 query latency             | 51.0 ms                    |
| p99 percentile query latency  | 91.0 ms                    |
```

#### After all changes - 1 client, 1 worker, websocket4modelcache

```
| Metric                        | Value                      |
|-------------------------------|----------------------------|
| Insertion time                | 33558 ms                   |
| Querying time                 | 40427 ms                   |
| Hit ratio                     | 1183 / 1668 (70.92326%)    |
| Expected query hit ratio      | 1092 / 1183 (92.30769%)    |
| Query throughput              | 41.259556 queries/second   |
| Mean query latency            | 24.215227 ms               |
| p95 query latency             | 26.0 ms                    |
| p99 percentile query latency  | 27.0 ms                    |
```
---
When using multiple workers we can see tremendous throughput gains (with the cost of some latency)

#### After all changes - 8 clients, 8 workers, websocket4modelcache
```
| Metric                        | Value                      |
|-------------------------------|----------------------------|
| Insertion time                | 12910 ms                   |
| Querying time                 | 17612 ms                   |
| Hit ratio                     | 1183 / 1668 (70.92326%)    |
| Expected query hit ratio      | 1092 / 1183 (92.30769%)    |
| Query throughput              | 94.70815 queries/second    |
| Mean query latency            | 84.247604 ms               |
| p95 query latency             | 110.0 ms                   |
| p99 percentile query latency  | 125.0 ms                   |
```

#### After all changes - 16 clients, 8 workers, websocket4modelcache
```
| Metric                        | Value                      |
|-------------------------------|----------------------------|
| Insertion time                | 13884 ms                   |
| Querying time                 | 15600 ms                   |
| Hit ratio                     | 1183 / 1668 (70.92326%)    |
| Expected query hit ratio      | 1092 / 1183 (92.30769%)    |
| Query throughput              | 106.92307 queries/second   |
| Mean query latency            | 148.90408 ms               |
| p95 query latency             | 178.65002 ms               |
| p99 percentile query latency  | 194.0 ms                   |
```
---
#### Contributors:
@Yuval-Roth 
@olgaoznovich 
@adiaybgu 
@omerdor001